+++ android/app/src/main/java/org/mudamos/petition/firebase/Config.java
@@ -40,13 +40,13 @@ public class Config {
     private void setDefaults() {
         Map<String, Object> defaults = new HashMap<>();
         defaults.put("authenticated_signers_button_title", "Lista de assinantes e outras informações");
-        defaults.put("plip_remaining_days_enabled", true);
-        defaults.put("link_get_to_know_mudamos", "https://www.mudamos.org");
-        defaults.put("link_help", "https://www.mudamos.org/ajuda");
-        defaults.put("link_send_your_idea", "https://www.mudamos.org/envie-sua-ideia");
+        defaults.put("plip_remaining_days_enabled", false);
+        defaults.put("link_get_to_know_mudamos", "https://www.mudamos.org/quem-somos");
+        defaults.put("link_help", "https://itsrio2.typeform.com/to/nGzwjv");
+        defaults.put("link_send_your_idea", "https://itsrio2.typeform.com/to/iulNZI");
         defaults.put("link_why_projects", "https://www.mudamos.org/institucional/projetos-de-lei-de-iniciativa-popular");
-        defaults.put("ineligible_to_sign_citywide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do município para o qual a lei se destina.");
-        defaults.put("ineligible_to_sign_statewide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do estado para o qual a lei se destina.");
+        defaults.put("ineligible_to_sign_citywide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do município para o qual ele se destina. Se deseja propor essa lei para o seu município, use a função \"Proponha um PL\" no menu do App.");
+        defaults.put("ineligible_to_sign_statewide_plip_reason", "Obrigado por seu apoio, mas você só pode assinar esse projeto de lei sendo eleitor do estado para o qual ele se destina. Se deseja propor essa lei para o seu município, use a função \"Proponha um PL\" no menu do App.");
 
         remoteConfig.setDefaults(defaults);
     }